myproject/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.1.0
Babel==2.10.3
backoff==2.2.1
basicsr==1.4.2
beautifulsoup4==4.12.2
bidict==0.23.0
blender.distutils==1.0.5
blinker==1.7.0
boto3==1.35.12
botocore==1.35.12
cachelib==0.12.0
cap-distutils-hack==0.0.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
Choco==1.0.5
click==8.1.7
cmake==3.25.2
collective.recipe.distutils==0.1
colorama==0.4.6
comtypes==1.2.0
contourpy==1.2.1
crispy-tailwind==0.5.0
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
decorator==5.1.1
distutils-extra-python==2.39.2
distutils-licenses==0.1.2
distutils-pytest==0.2.1
distutils-strtobool==0.1.0
distutils-ui==0.1.3
distutils644==0.4
Django==5.1
django-crispy-forms==2.1
django-filter==24.3
django-rest-authtoken==2.1.4
djangorestframework==3.15.0
dlib @ file:///C:/Users/MS/Desktop/Dlib_Windows_Python3.x-main/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
dnspython==2.6.1
docopt==0.6.2
docutils==0.20.1
drf-yasg==1.21.7
ebaysdk==2.1.5
einops==0.8.0
et-xmlfile==1.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
facexlib==0.3.0
fake-useragent==1.4.0
fastapi==0.108.0
ffmpy==0.4.0
filelock==3.15.4
filterpy==1.4.5
Flask==3.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Session==0.6.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
flux==1.3.5
flux-cli==0.1
fonttools==4.53.1
foris-plugins-distutils==0.31.1
freezegun==1.2.1
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
geoip2==2.9.0
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gradio==4.42.0
gradio_client==1.3.0
greenlet==3.0.3
grpcio==1.59.3
grpcio-tools==1.59.3
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==3.6
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
imutils==0.5.4
inflection==0.5.1
iniconfig==2.0.0
intel-openmp==2021.4.0
isodate==0.6.1
itsdangerous==2.1.2
jax==0.4.30
jaxlib==0.4.30
jdhp-distutils-demo==0.2
Jinja2==3.1.2
jmespath==1.0.1
jwt==1.3.1
keras==3.4.1
keras-nightly==3.4.1.dev2024071703
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
libsass==0.22.0
llvmlite==0.43.0
lmdb==1.5.1
loguru==0.7.2
lxml==5.2.1
lxml_html_clean==0.2.0
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
maxminddb==2.6.2
mdurl==0.1.2
mediapipe==0.10.14
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
mysql-connector-python==8.2.0
mysqlclient==2.2.4
namex==0.0.8
networkx==3.3
num2words==0.5.13
numba==0.60.0
numpy==1.26.2
ofxparse==0.21
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parsel==1.8.1
passlib==1.7.4
pdfminer.six==20240706
pillow==10.2.0
platformdirs==4.2.2
pluggy==1.3.0
polib==1.1.1
protobuf==4.21.12
protobuf-distutils==1.0
protobuf-grpc-distutils==0.1.5
psutil==5.9.8
psycopg2==2.9.9
py-cpuinfo==9.0.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==1.4.2
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==2.12.1
pypiwin32==223
pypng==0.20220715.0
pyqt-distutils==0.7.3
pyserial==3.5
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.1
python-stdnum==1.19
pyttsx3==2.90
pytz==2023.3.post1
pyusb==1.2.1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
realesrgan==0.3.0
reforis-distutils==0.1.1
regex==2024.7.24
reportlab==4.1.0
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
requests-viewer==0.0.12
requests2==2.16.0
requests3==0.0.0
rich==13.7.1
rjsmin==1.2.0
rl-renderPM==4.0.3
ruff==0.6.3
s3transfer==0.10.2
safetensors==0.4.4
scikit-image==0.24.0
scipy==1.14.0
scraperapi-sdk==0.2.2
scrapfly-sdk==0.8.10
seaborn==0.13.2
selenium==4.16.0
semantic-version==2.10.0
setuptools==71.1.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socks==0
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
starlette==0.32.0.post1
sympy==1.13.1
tb-nightly==2.18.0a20240717
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
tf-nightly==2.18.0.dev20240713
tf_keras==2.17.0
tf_nightly_intel==2.18.0.dev20240713
tifffile==2024.8.30
timm==1.0.9
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.44.2
trio==0.23.2
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.2.64
ultralytics-thop==2.0.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.25.0
vobject==0.9.6.1
w3lib==2.1.2
webdriver-manager==4.0.1
websockets==12.0
Werkzeug==3.0.1
wget==3.2
wheel==0.42.0
whitenoise==6.7.0
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
yapf==0.40.2
zc.buildout==3.0.1
zeep==4.2.1
zerokspot.recipe.distutils==0.1.2
zipp==3.20.1
